Create a GameRunner to handle switching between multiple screens and to clean up the main.rs file.
