+++ src/main.rs
@@ -8,22 +8,12 @@ mod save_management;
 mod movements;
 mod animations;
 mod items;
+mod game_runner;
 
 use macroquad::prelude::*;
-use game_state::GameState;
-use save_management::get_save_file_path;
-use ui::{NewGameMenu, render_death_screen};
-use ui::stat_display::stat_display;
-use ui::interaction_buttons::InteractionButton;
-use food::Food;
-use movements::get_sleeping_location;
+use game_runner::GameRunner;
 use utils::Location;
-use ui::play_area::draw_play_area;
-use shapes::sleeping_icon;
-use movements::{CreatureMovement, EggHop};
-use animations::creature_actions::{ActionAnimationType, CreatureActionAnimation};
 
-use ui::shop::ShopPage;
 
 pub const SCREEN_WIDTH: i32 = 200;
 pub const SCREEN_HEIGHT: i32 = 200;
@@ -33,81 +23,8 @@ pub const BACKGROUND_COLOR: Color = Color::new(0.8, 0.8, 0.8, 1.0);
 
 #[macroquad::main(main_window_conf)]
 async fn main() {
-    // Seed the random number generator
-    rand::srand(miniquad::date::now() as u64);
-    
-    let save_file_path = get_save_file_path();
-
-    let game_state = match GameState::from_file(&save_file_path).await {
-        Ok(mut state) => {
-            state.update();
-
-            if !state.creature().alive() {
-                render_death_screen(&state).await
-            } else {
-                state
-            }
-        },
-        Err(_) => NewGameMenu::default().render().await,
-    };
-
-    render_game(game_state).await;
-}
-
-async fn render_game(mut state: GameState) {
-    // Set some state
-    let buttons = InteractionButton::main_menu_buttons();
-    let mut sleeping_icon_movement = EggHop::new(get_sleeping_location(state.creature()).translate(-9.0, -16.0));
-    
-    let shop_btn = ShopPage::shop_button();
-
-    // Enter the actual game loop
-    loop {
-        state.update();
-
-        // If the creature has died, render the death screen and set the new state
-        if !state.creature().alive() {
-            state = render_death_screen(&state).await;
-        }
-
-        handle_button_click(&buttons, &mut state);
-        
-        clear_background(BACKGROUND_COLOR);
-        
-        // Draw the playing area the creature walks around in
-        draw_play_area(state.creature());
-        draw_creature(&mut state);
-
-        // Draw the "Zz" texture when sleeping
-        if state.creature().is_asleep() {
-            let location = sleeping_icon_movement.next_location();
-            draw_texture(&sleeping_icon(), location.x, location.y, WHITE);
-        }
-        
-        // Draw the creatures name and age
-        ui::draw_creature_name(&state);
-        ui::draw_age_display(&state);
-
-        for button in &buttons {
-            button.get_button().render();
-        }
-        
-        // If an animation is playing, render it
-        if let Some(animation) = state.current_animation.as_mut()
-            && animation.playing(){
-            animation.render();
-        }
-
-        if is_key_pressed(KeyCode::Escape) {
-            break;
-        }
-        
-        stat_display(state.creature());
-        shop_btn.render();
-        
-
-        next_frame().await;
-    }
+    let mut runner = GameRunner::initiate().await;
+    runner.render_game().await;
 }
 
 fn main_window_conf() -> Conf {
@@ -117,73 +34,5 @@ fn main_window_conf() -> Conf {
         window_height: SCREEN_HEIGHT,
         window_resizable: false,
         ..Default::default()
-    }
-}
-
-fn draw_creature(state: &mut GameState) {
-    // The creature shouldn't be drawn when an animation is playing.
-    if state.current_animation.is_some() {
-        return;
-    }
-
-    let creature_texture = state.creature().shape();
-    let creature_location = if state.creature().is_asleep() {
-        get_sleeping_location(state.creature())
-    } else {
-        state.creature_movement.next_location()
-    };
-
-    draw_texture_ex(
-        &creature_texture,
-        creature_location.x,
-        creature_location.y,
-        BLACK,
-        DrawTextureParams {
-            flip_x: state.creature_movement.mirror_sprite(),
-            ..Default::default()
-        }
-    );
-}
-
-fn handle_button_click(buttons: &[InteractionButton], game_state: &mut GameState) {
-    if game_state.current_animation.is_some() {
-        return;
-    }
-    
-    for button in buttons {
-        if button.get_button().is_clicked() {
-            match button {
-                InteractionButton::Energy(_) => game_state.creature_mut().toggle_sleep(),
-
-                InteractionButton::Food(_) => {
-                    let creature = game_state.creature_mut();
-                    if !creature.is_asleep() 
-                        && creature.food().value() != 100 
-                        && !creature.is_sick() 
-                    {
-                        let food = Food::new_random();
-                        creature.eat(food);
-                        game_state.set_animation(CreatureActionAnimation::new(ActionAnimationType::Eating(food)));
-                    }
-                },
-                InteractionButton::Joy(_) => {
-                    let creature = game_state.creature_mut();
-                    if !creature.is_asleep()
-                        && creature.joy().value() != 100
-                        && creature.energy().value() >= creature::PLAYING_ENERGY_COST
-                    {
-                        creature.play();
-                        game_state.set_animation(CreatureActionAnimation::new(ActionAnimationType::Play));
-                    }
-                },
-                InteractionButton::Health(_) => {
-                    let creature = game_state.creature_mut();
-                    if !creature.is_asleep() && creature.is_sick() {
-                        creature.heal();
-                        game_state.set_animation(CreatureActionAnimation::new(ActionAnimationType::Health));
-                    }
-                },
-            }
-        }
     }
 }
